backend/requirements.txt
django>=3.1.0, <3.2.0
django-widget-tweaks>=1.4.8, <1.5.0
django-registration>=3.2.0, <3.3.0
djangorestframework>=3.12.0, <3.13.0
djoser>=2.1.0, <2.2.0
psycopg2-binary
Pillow